sm changes

+++ app/Http/Controllers/PostsController.php
@@ -87,9 +87,13 @@ class PostsController extends Controller
     public function edit($id)
     {
         $post = BlogPost::findOrFail($id);
-        if (Gate::denies('update-post', $post)) {
-            abort(403, "You can't edit this blog post");
-        }
+//        if (Gate::denies('update-post', $post)) {
+//            abort(403, "You can't edit this blog post");
+//        }
+        //instead of gate
+        $this->authorize('update-post', $post);
+
+
         return view('posts.edit', ['post' => $post]);
     }
 
@@ -104,9 +108,11 @@ class PostsController extends Controller
     {
         $post = BlogPost::findOrFail($id);
 
-        if (Gate::denies('update-post', $post)) {
-            abort(403, "You can't edit this blog post");
-        }
+//        if (Gate::denies('update-post', $post)) {
+//            abort(403, "You can't edit this blog post");
+//        }
+        //instead of gate
+        $this->authorize('update-post', $post);
 
         $validated = $request->validated();
         $post->fill($validated);
@@ -142,6 +148,13 @@ class PostsController extends Controller
     public function destroy($id)
     {
         $post = BlogPost::findOrFail($id);
+
+//        if (Gate::denies('delete-post', $post)) {
+//            abort(403, "You can't delete this blog post");
+//        }
+        //instead of gate
+        $this->authorize('delete-post', $post);
+
         $post->delete();
 
         session()->flash('status', 'Blog post was deleted!');
